Support executing from pipes and other streams
This is done by reading into a string builder. The whole stream
must be read and EOF must occur before compilation will happen.

This is obviously not ideal, but supporting incremental or
on-the-fly compilation would require a fairly extensive rewrite
of the scanner and compiler, which is not possible at the moment.

At the very least, this allows interesting cases like passing a
Bash dynamic redirect (where a named pipe is created and passed
as an argument) to the interpreter as a file to execute, which is
particularly useful for things like Bim to pass configuration
scripts on the command line.